Expose bundle-init wizard in empty workspace
- Moved out bundle init logic out of BundleProjectManager into BundleInitWizard
- Moved some logic to "pure" functions so they can be called without
  initializing a bunch of dependencies
- Welcome view in the empty workspace now has Initialize Project button
