+++ packages/databricks-vscode/src/bundle/BundleProjectManager.ts
@@ -1,18 +1,9 @@
-import {
-    ExtensionContext,
-    QuickPickItem,
-    QuickPickItemKind,
-    Disposable,
-    Uri,
-    window,
-    commands,
-    TerminalLocation,
-} from "vscode";
+import {ExtensionContext, Disposable, Uri, window} from "vscode";
 import fs from "node:fs/promises";
 import path from "path";
 import {ConnectionManager} from "../configuration/ConnectionManager";
 import {ConfigModel} from "../configuration/models/ConfigModel";
-import {BundleFileSet} from "./BundleFileSet";
+import {BundleFileSet, getSubProjects} from "./BundleFileSet";
 import {logging} from "@databricks/databricks-sdk";
 import {Loggers} from "../logger";
 import {CachedValue} from "../locking/CachedValue";
@@ -20,13 +11,11 @@ import {CustomWhenContext} from "../vscode-objs/CustomWhenContext";
 import {CliWrapper} from "../cli/CliWrapper";
 import {LoginWizard, saveNewProfile} from "../configuration/LoginWizard";
 import {Mutex} from "../locking";
-import {
-    AuthProvider,
-    ProfileAuthProvider,
-} from "../configuration/auth/AuthProvider";
+import {ProfileAuthProvider} from "../configuration/auth/AuthProvider";
 import {ProjectConfigFile} from "../file-managers/ProjectConfigFile";
 import {randomUUID} from "crypto";
 import {onError} from "../utils/onErrorDecorator";
+import {BundleInitWizard, promptToOpenSubProjects} from "./BundleInitWizard";
 
 export class BundleProjectManager {
     private logger = logging.NamedLogger.getOrCreate(Loggers.Extension);
@@ -134,7 +123,7 @@ export class BundleProjectManager {
     }
 
     private async detectSubProjects() {
-        this.subProjects = await this.bundleFileSet.getSubProjects();
+        this.subProjects = await getSubProjects(this.workspaceUri);
         this.logger.debug(
             `Detected ${this.subProjects?.length} sub folders with bundle projects`
         );
@@ -145,36 +134,8 @@ export class BundleProjectManager {
 
     public async openSubProjects() {
         if (this.subProjects && this.subProjects.length > 0) {
-            return this.promptToOpenSubProjects(this.subProjects);
+            return promptToOpenSubProjects(this.subProjects);
         }
-    }
-
-    private async promptToOpenSubProjects(
-        projects: {absolute: Uri; relative: Uri}[]
-    ) {
-        type OpenProjectItem = QuickPickItem & {uri?: Uri};
-        const items: OpenProjectItem[] = projects.map((project) => {
-            return {
-                uri: project.absolute,
-                label: project.relative.fsPath,
-                detail: project.absolute.fsPath,
-            };
-        });
-        items.push(
-            {label: "", kind: QuickPickItemKind.Separator},
-            {label: "Choose another folder"}
-        );
-        const options = {
-            title: "Select the project you want to open",
-        };
-        const item = await window.showQuickPick<OpenProjectItem>(
-            items,
-            options
-        );
-        if (!item) {
-            return;
-        }
-        await commands.executeCommand("vscode.openFolder", item.uri);
     }
 
     private async detectLegacyProjectConfig() {
@@ -291,171 +252,17 @@ export class BundleProjectManager {
     }
 
     public async initNewProject() {
-        const authProvider = await this.configureAuthForBundleInit();
-        if (!authProvider) {
-            this.logger.debug(
-                "No valid auth providers, can't proceed with bundle init wizard"
-            );
-            return;
-        }
-        const parentFolder = await this.promptForParentFolder();
-        if (!parentFolder) {
-            this.logger.debug("No parent folder provided");
-            return;
-        }
-        await this.bundleInitInTerminal(parentFolder, authProvider);
-        this.logger.debug(
-            "Finished bundle init wizard, detecting projects to initialize or open"
-        );
-        await this.isBundleProjectCache.refresh();
-        const projects = await this.bundleFileSet.getSubProjects(parentFolder);
-        if (projects.length > 0) {
-            this.logger.debug(
-                `Detected ${projects.length} sub projects after the init wizard, prompting to open one`
-            );
-            await this.promptToOpenSubProjects(projects);
-        } else {
-            this.logger.debug(
-                `No projects detected after the init wizard, showing notification to open a folder manually`
-            );
-            const choice = await window.showInformationMessage(
-                `We haven't detected any Databricks projects in "${parentFolder.fsPath}". If you initialized your project somewhere else, please open the folder manually.`,
-                "Open Folder"
-            );
-            if (choice === "Open Folder") {
-                await commands.executeCommand("vscode.openFolder");
-            }
-        }
-    }
-
-    private async configureAuthForBundleInit(): Promise<
-        AuthProvider | undefined
-    > {
-        let authProvider =
+        const bundleInitWizard = new BundleInitWizard(this.cli);
+        const authProvider =
             this.connectionManager.databricksWorkspace?.authProvider;
-        if (authProvider) {
-            const response = await this.promptToUseExistingAuth(authProvider);
-            if (response.cancelled) {
-                return undefined;
-            } else if (!response.approved) {
-                authProvider = undefined;
-            }
-        }
-        if (!authProvider) {
-            authProvider = await LoginWizard.run(this.cli, this.configModel);
-        }
-        if (authProvider && (await authProvider.check())) {
-            return authProvider;
-        } else {
-            return undefined;
-        }
-    }
-
-    private async promptToUseExistingAuth(authProvider: AuthProvider) {
-        type AuthSelectionItem = QuickPickItem & {approved: boolean};
-        const items: AuthSelectionItem[] = [
-            {
-                label: "Use current auth",
-                detail: `Host: ${authProvider.host.hostname}`,
-                approved: true,
-            },
-            {
-                label: "Setup new auth",
-                approved: false,
-            },
-        ];
-        const options = {
-            title: "What auth do you want to use for the new project?",
-        };
-        const item = await window.showQuickPick<AuthSelectionItem>(
-            items,
-            options
+        const parentFolder = await bundleInitWizard.initNewProject(
+            this.workspaceUri,
+            authProvider,
+            this.configModel
         );
-        return {
-            cancelled: item === undefined,
-            approved: item?.approved ?? false,
-        };
-    }
-
-    private async bundleInitInTerminal(
-        parentFolder: Uri,
-        authProvider: AuthProvider
-    ) {
-        const terminal = window.createTerminal({
-            name: "Databricks Project Init",
-            isTransient: true,
-            location: TerminalLocation.Editor,
-            env: this.cli.getBundleInitEnvVars(authProvider),
-        });
-        const args = [
-            "bundle",
-            "init",
-            "--output-dir",
-            this.cli.escapePathArgument(parentFolder.fsPath),
-        ];
-        const finalPrompt = `echo "Press any key to close the terminal and continue ..."; read; exit`;
-        terminal.sendText(
-            `${this.cli.cliPath} ${args.join(" ")}; ${finalPrompt}`
-        );
-        return new Promise<void>((resolve) => {
-            const closeEvent = window.onDidCloseTerminal(async (t) => {
-                if (t !== terminal) {
-                    return;
-                }
-                closeEvent.dispose();
-                resolve();
-            });
-            this.disposables.push(closeEvent);
-        });
-    }
-
-    private async promptForParentFolder(): Promise<Uri | undefined> {
-        const quickPick = window.createQuickPick();
-        const openFolderLabel = "Open folder selection dialog";
-        quickPick.value = this.workspaceUri.fsPath;
-        quickPick.title =
-            "Provide a path to a folder where you would want your new project to be";
-        quickPick.items = [
-            {label: quickPick.value, alwaysShow: true},
-            {label: "", kind: QuickPickItemKind.Separator, alwaysShow: true},
-            {label: openFolderLabel, alwaysShow: true},
-        ];
-        quickPick.show();
-        const disposables = [
-            quickPick.onDidChangeValue(() => {
-                quickPick.items = [
-                    {label: quickPick.value, alwaysShow: true},
-                    ...quickPick.items.slice(1),
-                ];
-            }),
-        ];
-        const choice = await new Promise<QuickPickItem | undefined>(
-            (resolve) => {
-                disposables.push(
-                    quickPick.onDidAccept(() =>
-                        resolve(quickPick.selectedItems[0])
-                    ),
-                    quickPick.onDidHide(() => resolve(undefined))
-                );
-            }
-        );
-        disposables.forEach((d) => d.dispose());
-        quickPick.hide();
-        if (!choice) {
-            return;
+        if (parentFolder) {
+            await this.isBundleProjectCache.refresh();
         }
-        if (choice.label !== openFolderLabel) {
-            return Uri.file(choice.label);
-        }
-        const choices = await window.showOpenDialog({
-            title: "Chose a folder where you would want your new project to be",
-            openLabel: "Select folder",
-            defaultUri: this.workspaceUri,
-            canSelectFolders: true,
-            canSelectFiles: false,
-            canSelectMany: false,
-        });
-        return choices ? choices[0] : undefined;
     }
 
     dispose() {
